TRESTLE-599: Build fixes, also removed short from the TDTree tuples, to help make maths quicker. #time 15m

+++ trestle-reasoner/src/main/java/com/nickrobison/trestle/reasoner/caching/tdtree/LeafNode.java
@@ -44,7 +44,7 @@ public abstract class LeafNode<Value> {
 
     public double[] getLeafVerticies() {
         return TDTreeHelpers.getTriangleVerticies(TDTreeHelpers.adjustedLength[getIDLength(this.leafID)],
-                this.leafMetadata.getShort(3),
+                this.leafMetadata.getInt(3),
                 this.leafMetadata.getDouble(1),
                 this.leafMetadata.getDouble(2));
     }
